Sprint 2 - Menu code restructuring and Sprint 3 - updating orders information
