+++ cart.html
@@ -16,15 +16,6 @@
             <tbody id="viewCart_tbody">
 
             </tbody>
-            <!-- <tr>
-                <td class="itemName"">item 1</td>
-                <td class="itemPrice">item 1</td>
-                <td class="itemQuantity">item 1</td>
-            </tr>
-            <tr>
-                <td colspan="2" class="itemTotalPrice">Total Price:</td>
-                <td class="totalprice"></td>
-            </tr> -->
         </table>
         <input type="submit" value="Make payment" class="cartPayment">
         <br><br>
