+++ index.html
@@ -50,7 +50,6 @@
 </body>
 <footer>
     <script type="text/javascript" src="js/include.js"></script>
-    <script type="text/javascript" src="js/menu.js"></script>
     <script src="https://kit.fontawesome.com/09cf2947bd.js" crossorigin="anonymous"></script>
     <script>
         includeHTML();
