+++ js/main.js
@@ -1,12 +1,9 @@
 import { LoginUI } from "./boundary/loginUI.js";
 import { AdminUI } from "./boundary/adminUI.js";
 import { MenuUI } from "./boundary/menuUI.js";
+import { CartUI } from "./boundary/cartUI.js";
 import { OrdersUI } from "./boundary/ordersUI.js";
 import * as Validation from "./validation.js";
-
-// Global config
-let MAX_ORDER_QUANTITY = 10;
-
 
 // Add event listener for login click button
 const loginForm = document.getElementById("login_form");
@@ -208,93 +205,14 @@ profileUpdateForm.addEventListener("submit", async function (e) {
 
 
 // ========== Menu functions ==========
-const menu = document.getElementById("defaultOpen");
-menu.addEventListener("click", async function (e) {
+const menuPage = document.getElementById("defaultOpen");
+menuPage.addEventListener("click", async function (e) {
     e.preventDefault();
-
-    displayMenuTable();
+    const menu = new MenuUI();
+    menu.displayMenu();
 });
 // Run this by default when page load to display menu
 document.getElementById("defaultOpen").click();
-
-async function displayMenuTable() {
-    const menuTable = document.getElementById("menu_table");
-    menuTable.innerHTML = "";
-
-    const menu = new MenuUI();
-    let categories = await menu.getMenuCat();
-    let menuItems = await menu.getMenuItems();
-
-    sessionStorage.setItem("menuItem", JSON.stringify(menuItems))
-
-    categories.forEach((cat) => {
-        // console.log(cat);
-
-        const catHead = menuTable.insertRow();
-        let th = document.createElement("th");
-        th.setAttribute("class", "categoryTitle");
-        th.setAttribute("colspan", 3);
-        th.innerHTML = cat.catName;
-        catHead.appendChild(th);
-
-        let columnCount = 0;
-        let tr;
-        menuItems.forEach((items) => {
-            if (cat.catID === items.itemCategory) {
-                if (columnCount == 0) {
-                    tr = menuTable.insertRow();
-                }     
-                const itemTd = tr.insertCell();
-                
-                // Add item image
-                let menuImage = document.createElement("div");
-                menuImage.setAttribute("class", "itemImage");
-                menuImage.innerHTML = items.itemImage;
-                itemTd.appendChild(menuImage);
-
-                // Add item name
-                let menuName = document.createElement("div");
-                menuName.setAttribute("class", "itemName");
-                menuName.innerHTML = items.itemName;
-                itemTd.appendChild(menuName);
-
-                // Add item price
-                let menuPrice = document.createElement("div");
-                menuPrice.setAttribute("class", "itemPrice");
-                menuPrice.innerHTML = "$" + items.itemPrice.toFixed(2);
-                itemTd.appendChild(menuPrice);
-
-                columnCount += 1;
-                if (columnCount == 3)
-                    columnCount = 0;
-
-                // Temporary listener to add item to cart
-                itemTd.addEventListener("click", function (e) {
-                    openOverlay(items);
-                    const addToCartBtn = document.getElementById("overlay_addToCart");
-                    addToCartBtn.addEventListener("click", function (e) {
-                        e.preventDefault();
-                        let cartOrder = (sessionStorage.getItem("cartOrder") != null) ? JSON.parse(sessionStorage.getItem("cartOrder")) : {};
-                        
-                        let quantityOrdered = Number(document.getElementById("overlay_quantity").value);
-                        if (items.id in cartOrder) {
-                            if ((cartOrder[items.id] + quantityOrdered) > MAX_ORDER_QUANTITY)
-                                cartOrder[items.id] = MAX_ORDER_QUANTITY;
-                            else
-                                cartOrder[items.id] = cartOrder[items.id] + quantityOrdered;
-                        }
-                        else {
-                            cartOrder[items.id] = quantityOrdered;
-                        }
-
-                        sessionStorage.setItem("cartOrder", JSON.stringify(cartOrder));
-                        document.getElementById("overlay_closeBtn").click();
-                    }, {once : true});
-                });
-            }
-        });
-    });
-}
 
 // Close overlay
 const closeOverlayBtn = document.getElementById("overlay_closeBtn");
@@ -304,24 +222,19 @@ closeOverlayBtn.addEventListener("click", function (e) {
     document.getElementById("overlay_quantity").value = 1;
 });
 
-function openOverlay(items) {
-    document.getElementById("overlay_itemImage").innerHTML = items.itemImage;
-    document.getElementById("overlay_itemName").innerHTML = items.itemName;
-    document.getElementById("overlay_itemPrice").innerHTML = "$" + items.itemPrice.toFixed(2);
-    document.getElementById("menuItem_overlay").style.height = "100%";
-}
-
 const overlayMinusBtn = document.getElementById("overlay_minus");
 overlayMinusBtn.addEventListener("click", function(e) {
     e.preventDefault();
-    minus("overlay_quantity");
+    const menu = new MenuUI();
+    menu.minus("overlay_quantity");
 });
+
 const overlayPlusBtn = document.getElementById("overlay_plus");
 overlayPlusBtn.addEventListener("click", function(e) {
     e.preventDefault();
-    plus("overlay_quantity");
+    const menu = new MenuUI();
+    menu.plus("overlay_quantity");
 });
-
 // ====== End of Menu functions =======
 
 
@@ -330,7 +243,8 @@ overlayPlusBtn.addEventListener("click", function(e) {
 const cartBtn = document.getElementById("cartTab");
 cartBtn.addEventListener("click", async function(e) {
     e.preventDefault();
-    displayCart();
+    const cart = new CartUI();
+    cart.displayCart();
 });
 
 const cartForm = document.getElementById("cart_form");
@@ -349,121 +263,6 @@ cartForm.addEventListener("submit", async function (e) {
     }
 });
 
-// Display cart function
-function displayCart() {
-    const cartTable = document.getElementById("viewCart_tbody");
-    cartTable.innerHTML = "";
-    const cartOrder = JSON.parse(sessionStorage.getItem("cartOrder"));
-    const menuItem = JSON.parse(sessionStorage.getItem("menuItem"));
-
-    let totalPrice = 0;
-    if (JSON.parse(sessionStorage.getItem("cartOrder")) != null) {
-        for (const [key, value] of Object.entries(cartOrder)) {
-            let item;
-            for (const itm of menuItem) {
-                if (itm.id === key) {
-                    item = itm;
-                    break;
-                }
-            }
-            const itemRow = cartTable.insertRow();
-            const nameTd = itemRow.insertCell();
-            nameTd.setAttribute("class", "itemName");
-            nameTd.innerHTML = item.itemName;
-    
-            const priceTd = itemRow.insertCell();
-            priceTd.setAttribute("class", "itemPrice");
-            priceTd.innerHTML = "$" + item.itemPrice.toFixed(2);
-    
-            const quantityTd = itemRow.insertCell();
-            quantityTd.setAttribute("class", "itemQuantity");
-            
-            const minusDiv = document.createElement("span");
-            minusDiv.setAttribute("class", "cartMinus");
-            minusDiv.setAttribute("id", item.id);
-            minusDiv.innerHTML = "&minus;";
-            
-            const quantityDiv = document.createElement("input");
-            quantityDiv.setAttribute("class", "cartQuantity");
-            quantityDiv.setAttribute("id", "cartQuantity" + item.id);
-            quantityDiv.setAttribute("name", "cartQuantity");
-            quantityDiv.setAttribute("value", value);
-            quantityDiv.readOnly = true;
-
-            const plusDiv = document.createElement("span");
-            plusDiv.setAttribute("class", "cartPlus");
-            plusDiv.setAttribute("id", item.id);
-            plusDiv.innerHTML = "&#43;";
-            
-            quantityTd.appendChild(minusDiv);
-            quantityTd.appendChild(quantityDiv);
-            quantityTd.appendChild(plusDiv);
-    
-            const removeTd = itemRow.insertCell();
-            removeTd.setAttribute("class", "itemRemove");
-            removeTd.setAttribute("id", item.id);
-            removeTd.innerHTML = '<i class="fa-solid fa-xmark"></i>';
-    
-            totalPrice += item.itemPrice * value;
-        }
-    
-        const totalPriceRow = cartTable.insertRow();
-        const priceLabel = totalPriceRow.insertCell();
-        priceLabel.setAttribute("class", "itemTotalPriceLabel");
-        priceLabel.innerHTML = "Total Price:";
-    
-        const totalPriceTd = totalPriceRow.insertCell();
-        totalPriceTd.setAttribute("class", "totalPrice");
-        totalPriceTd.innerHTML = "$" + totalPrice.toFixed(2);
-        const totalPriceInput = document.createElement("input");
-        totalPriceInput.setAttribute("name", "totalPrice");
-        totalPriceInput.setAttribute("value", totalPrice.toFixed(2));
-        totalPriceInput.setAttribute("style", "display:none");
-        totalPriceTd.appendChild(totalPriceInput);
-    
-        // Remove item from cart
-        const removeBtn = document.getElementsByClassName("itemRemove");
-        Array.from(removeBtn).forEach((btn) => {
-            btn.addEventListener("click", function (e) {
-                e.preventDefault();
-                delete cartOrder[btn.id];
-                sessionStorage.setItem("cartOrder", JSON.stringify(cartOrder));
-                displayCart();
-            })
-        });
-    }
-
-    const cartMinusBtn = document.getElementsByClassName("cartMinus");
-    Array.from(cartMinusBtn).forEach((btn) => {
-        btn.addEventListener("click", function(e) {
-            e.preventDefault();
-            console.log(btn.id);
-            minus("cartQuantity", "cart", btn.id);
-            updateCart(btn.id);
-        });
-    });
-    const cartPlusBtn = document.getElementsByClassName("cartPlus");
-    Array.from(cartPlusBtn).forEach((btn) => {
-        btn.addEventListener("click", function(e) {
-            e.preventDefault();
-            console.log(btn.id);
-            plus("cartQuantity", "cart", btn.id);
-            updateCart(btn.id);
-        });
-    });
-}
-
-function updateCart(itemId) {
-    let cart = JSON.parse(sessionStorage.getItem("cartOrder"));
-    
-    if (Number(document.getElementById("cartQuantity" + itemId).value) != 0) {
-        cart[itemId] = Number(document.getElementById("cartQuantity" + itemId).value);
-        sessionStorage.setItem("cartOrder", JSON.stringify(cart));
-    }
-
-    displayCart();
-}
-
 // ====== End of Cart functions =======
 
 
@@ -481,55 +280,10 @@ paymentForm.addEventListener("submit", async function (e) {
         document.location.href = "./index.html";
         alert("Payment success");
     }
-        
-    // TODO: Include a popup that say payment successful
 });
 
 // ===== End of Payment functions =====
 
-
-// ====== Time tracking function ======
-// !!!!DO NOT TOUCH!!!!
-window.timeSpent = 0;
-window.timer = setInterval(function () {
-    timeSpent += 1;
-    // console.log("Timespent : " + timeSpent);
-}, 1000);
-// ==== End of trackiong functions ====
-
-// ======= Plus and minus function to add quantity =======
-function minus(quantityId, form = "overlay", itemId = "") {
-    let quantity;
-    let id = (form === "overlay") ? quantityId : quantityId + itemId;
-
-    quantity = Number(document.getElementById(id).value) - 1;
-
-    if (quantity < 1 && form === "overlay") {
-        quantity = 1;
-    }
-    else {
-        if (quantity == 0) {
-            let cart = JSON.parse(sessionStorage.getItem("cartOrder"));
-            delete cart[itemId];
-            sessionStorage.setItem("cartOrder", JSON.stringify(cart));
-        }
-    }
-    document.getElementById(id).value = quantity;
-   
-}
-
-function plus(quantityId, form = "overlay", itemId = "") { // need to have a config setting for quantity limit
-    let quantity;
-    let id = (form === "overlay") ? quantityId : quantityId + itemId;
-
-    quantity = Number(document.getElementById(id).value) + 1;
-
-    if (quantity > MAX_ORDER_QUANTITY)
-        quantity = MAX_ORDER_QUANTITY;
-
-    document.getElementById(id).value = quantity;
-}
-// =======================================================
 
 // ========== Manage Order functions ==========
 
@@ -582,3 +336,24 @@ orderUpdateForm.addEventListener("change", (event) => {
     ticketButton.style.display = closeOption.value == 2 ? 'block' : 'none';        
 });    
 // ====== End of Manage Order functions =======
+
+
+
+
+
+
+
+
+
+
+
+
+// ====== Time tracking function ======
+// ALWAYS MAKE SURE ITS AT THE BOTTOM OF THE CODES
+// !!!!DO NOT TOUCH!!!!
+window.timeSpent = 0;
+window.timer = setInterval(function () {
+    timeSpent += 1;
+    // console.log("Timespent : " + timeSpent);
+}, 1000);
+// ==== End of trackiong functions ====
